rolling back fin_utils, worker interface, session changes
